Entrenador_golf/requirements.txt
streamlit
google-genai
youtube-transcript-api
requests
langdetect
Pillow
pdfplumber
streamlit-calendar
